feat(refactor-mcts): Separate InvalidChoice class from ChooseFromCardIDs
- Add comments for doxygen
